fix: Refactor RequestUtil package structure
