+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/actors/QuestionSetActor.scala
@@ -14,7 +14,7 @@ import org.sunbird.graph.dac.model.Node
 import org.sunbird.graph.nodes.DataNode
 import org.sunbird.managers.questionset.HierarchyManager.hierarchyPrefix
 import org.sunbird.managers.questionset.{AssessmentManager, CopyManager, HierarchyManager, UpdateHierarchyManager}
-import org.sunbird.utils.RequestUtil
+import org.sunbird.utils.questionset.RequestUtil
 
 import java.util
 import javax.inject.Inject
