+++ assessment-api/assessment-actors/src/main/scala/org/sunbird/v5/actors/QuestionSetActor.scala
@@ -15,7 +15,8 @@ import org.sunbird.graph.nodes.DataNode
 import org.sunbird.graph.schema.DefinitionNode
 import org.sunbird.managers.questionset.HierarchyManager.hierarchyPrefix
 import org.sunbird.managers.questionset.{CopyManager, HierarchyManager, UpdateHierarchyManager}
-import org.sunbird.utils.{AssessmentErrorCodes, RequestUtil}
+import org.sunbird.utils.AssessmentErrorCodes
+import org.sunbird.utils.questionset.RequestUtil
 import org.sunbird.v5.managers.AssessmentV5Manager
 
 import java.util
