+++ taxonomy-api/taxonomy-actors/src/main/scala/org/sunbird/utils/RequestUtil.scala
@@ -1,4 +1,4 @@
-package org.sunbird.utils
+package org.sunbird.utils.taxonomy
 
 import org.sunbird.common.dto.Request
 import org.sunbird.common.exception.ClientException
